requirements.txt
# Milestone 2 Requirements
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
requests>=2.31.0
numpy>=1.24.0